增加了账户管理模块待测试,增加了一个dto映射工具类

+++ src/main/java/org/mxwj/librarymanagement/model/Account.java
@@ -26,7 +26,7 @@ public class Account {
     private String email;
     
     @Column(name = "user_type")
-    private Short userType;     // 0:普通用户, 1:管理员
+    private Integer userType;     // 0:普通用户, 1:管理员
 
     @Column(name = "created_at")
     private OffsetDateTime createdAt;
@@ -34,5 +34,5 @@ public class Account {
     @Column(name = "last_login")
     private OffsetDateTime lastLogin;
     
-    private Short status;  // 1:正常, 0:禁用
+    private Integer status;  // 1:正常, 0:禁用
 }
